+++ shell/run.e2e.tests.sh
@@ -4,6 +4,11 @@ source ./pre_run.sh
 
 printf "\nStarting end-to-end tests\n\n"
 
+# We had to source the nvm script to solve the following error:
+#.  nvm is not compatible with the npm config "prefix" option: currently set to ""
+#.  Run `npm config delete prefix` or `nvm use --delete-prefix v23.9.0 --silent` to unset it.
+source "${HOME}/.nvm/nvm.sh"
+
 npm run cy:test
 
 APP_PORT=5010